+++ src/common/libkvs/kvs_txn.h
@@ -43,7 +43,7 @@ int flux_kvs_txn_put_raw (flux_kvs_txn_t *txn,
                           int flags,
                           const char *key,
                           const void *data,
-                          int len);
+                          size_t len);
 
 int flux_kvs_txn_put_treeobj (flux_kvs_txn_t *txn,
                               int flags,
